+++ src/docs/examples/Link/Example.tsx
@@ -1,0 +1,10 @@
+import { Link } from 'uilib';
+
+export default () => {
+  return (
+    <>
+      <Link href="/button">local link</Link>
+      <Link href="https://apostol.space">external link</Link>
+    </>
+  );
+};