+++ src/docs/examples/Link/index.tsx
@@ -1,0 +1,5 @@
+import { Code } from 'docs/components';
+
+import example from '!!raw-loader!./Example';
+
+export default () => <Code code={example} />;